[#16] Add refinement for flexible hash access
- Introduced `FlexibleHashAccess` refinement to allow hash key access using both strings and symbols interchangeably.
- Moved `LoggerTraceRefinement` from logging.rb
- Updated `lib/familia.rb` to require the new refinements.
